Requests mapping for tournament

+++ src/main/java/com/projet/Tournament/Tournament.java
@@ -20,25 +20,25 @@ public class Tournament {
             generator = "tournament_sequence"
 
     )
-    private Integer  id;
+    private Long  id;
     private String name;
     private boolean isPrivate;
 
     public Tournament() {
     }
 
-    public Tournament (Integer id, String name,
+    public Tournament (Long id, String name,
                       boolean isPrivate) {
         this.name = name;
         this.id = id;
         this.isPrivate = isPrivate;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
